view.h was refusing to compile so I added in namespace std; Was it compiling for you guys before?
Added a while loop in controller.cpp to check if player number input is valid.
Before it allowed the program to continue and error out.

Added a stub to start off on checking adjacent cities in Ly's new function format

Added a new variable in playerchar struct :  ActionsLeft
I'm going to try an approach in which we keep the player's number of actions remaining in an int.

This way, we can account for events and profession bonuses.

Ex we can set the int to give the player 6 moves instead of 5, this will be added at the start of every turn.

A while loop will run until ActionsLeft hits 0.
Let me know if you need clarification