Finished Saved/Load Functions
They work. The only issue is ; when saving a game, the name of the game should not include any spaces or '#'s . I am not very good with string functions so I wasn't able to come up with a way to removes white spaces from a string. So the name has to be a single word without spaces.

Everything else works.