Had issues with merging model.h so I ended up just rewriting parts of my code.
PlayEventCard is now completely separate from PlayCard
If you want to play an event card alone, just insert the player's number and
the card itself (no indexes or values). The function will play the card, then look up the card's info in the given player's hand and discard it.

The function called for discarding it is discardCard(playernum, card);
The function takes in a card, looks for it in the player's hand and discards it.

Note to self: Add in an output to state the card that was discarded.


One of Darshan's code was causing the program to error when trying to compile.

Line 1557 on model.h, you put in PlayerDeckD.size < 1
I forget how that function worked, but I changed it to PlayerDeckD.size() < 1

That seemed to fix it, you might want to take a doublecheck on it to see if that messed anything up.