Operations Expert Implemented.
Commented out system("CLR"), it was kind of in the way when debugging
I think we could find a better way to use system("CLR") instead of just sticking system pause everywhere, I'll see if I can figure it out later

Forgot to put x = 1 in the while loop for shuttleFlight. Fixed that so if you try to move to the same location nothing happens and you get your turn back. Thanks for pointing that out Ly